computer_vision/processor/requirements.txt
opencv-python
numpy
ultralytics
deep_sort_realtime
azure-eventhub
azure-servicebus
python-dotenv
azure-identity
azure-appconfiguration
azure-keyvault-secrets
scipy